WEB-2039 extract actions list sheet
