+++ src/sheet-common.css.ts
@@ -256,38 +256,3 @@ export const bodyContent = style({
         },
     },
 });
-
-export const sheetActionRow = style([
-    sprinkles({
-        display: 'flex',
-        padding: 16,
-        minHeight: 72,
-        alignItems: 'center',
-    }),
-    {
-        transition: 'background-color 0.1s ease-in-out',
-        ':active': {
-            background: skinVars.colors.backgroundContainerPressed,
-        },
-        '@media': {
-            [mq.supportsHover]: {
-                ':hover': {
-                    background: skinVars.colors.backgroundContainerHover,
-                },
-                // need to repeat this inside of @media to avoid :hover background to take precedence over :active
-                ':active': {
-                    background: skinVars.colors.backgroundContainerPressed,
-                },
-            },
-            [mq.touchableOnly]: {
-                transition: 'none',
-            },
-        },
-    },
-]);
-
-export const infoItemIcon = sprinkles({
-    display: 'flex',
-    alignItems: 'center',
-    height: 24,
-});
