+++ src/sheet-info.css.ts
@@ -1,0 +1,7 @@
+import {sprinkles} from './sprinkles.css';
+
+export const infoItemIcon = sprinkles({
+    display: 'flex',
+    alignItems: 'center',
+    height: 24,
+});
